pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "healthcraft"
version = "0.1.0"
description = "Emergency Medicine RL Training Environment"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.9.0",
]
mcp = [
    "mcp>=1.0",
    "uvicorn>=0.30",
]
openem = [
    "openem>=0.2.0",
]
docker = [
    "psycopg2-binary>=2.9",
]
all = [
    "healthcraft[dev,mcp,openem,docker]",
]

[project.scripts]
healthcraft = "healthcraft.cli:main"

[project.urls]
Homepage = "https://github.com/GOATnote-Inc/healthcraft"
Documentation = "https://github.com/GOATnote-Inc/healthcraft/tree/main/docs"
Repository = "https://github.com/GOATnote-Inc/healthcraft"
Issues = "https://github.com/GOATnote-Inc/healthcraft/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
